Improved R Initialization and Library Loading
The method for executing unsafe code (which might throw exceptions)
has been improved. It now returns TRUE on success and FALSE on failure.
This  allows us to have a better structured control flow.

In order to load libraries on the fly, on unixoid systems, you need a
user library as the system-wide R library requires you need sudo or
similar permissions. Whatever. Well, you can supply a path to a user
library via the environment variable R_LIBS_USER. So we do this. We
first check some default folder locations and, if they do not exist,
create a folder in the /tmp/ path. Although this fallback will leave
behind a temporary folder when our programs terminate, this won't do
much harm. The folder will be deleted upon reboot (remember, unixoid).