First Attempt for Better 2D Matrix Iterations
This is the first attempt to build a more versatile and robust way to
iterate over multiple matrices in two dimensions. One dimension is
selected as "x" dimension, the other one is the "y" dimension. For each
unique value along the "x" dimension, we construct a matrix of all
corresponding "y" values from all the matrices. The iteration is now
based on a visitor pattern instead of an iterator pattern. Thus, it is
more directly suitable for parallelization.