InterQuantileRange now always set to infinity if invinity involved
If either of the two quantiles involved in the inter-quantile range
computation is infinite, the inter-quantile range becomes infinite too.
This is even the case if both quantiles are, say, negative infinity.
This makes sense as the inter-quantile range would also be infinite if
smaller quantile was negative infinite and the larger one would be a
very small number.