Added Policy for Dealing with NaNs to MatrixIteration2D
We now have a policy for dealing with NaN values during 2D matrix
iterations: Leading and trailing NaNs can be skipped both along the
x and y axis. NaN values occuring somewhere in between at the y axis
may be replaced with a constant. For the x-axis, this is not allowed.
Any NaN not replaced or skipped will lead to an IllegalStateException.