Added Implementation of new Statistic Information Interface + Builder
The new IStatisticInformation interface defined in utils-base has been
implemented in form of an immutable object. A builder for constructing
such an object has been provided. The builder implements IAggregate,
i.e., allows the user to add an arbitrary number of values from which
internal aggregates compute the values of the information record.