New QuickStableSum + StableSum Better Deals with Longs
The new QuickStableSum is faster than StableSum and suitable for
situations where we add up values to a double result, i.e., do not
need to worry about integer arithmetic and do not care about the
difference between overflow, infinity, and NaN too much.

Adding of long values to sums has been improved: It is now a two-step
process which first adds the double version of the long and then a
potential residue resulting from the double conversion.